chore: split up repo files
