+++ internal/database/component.go
@@ -1,0 +1,46 @@
+package database
+
+import (
+	"context"
+	"fmt"
+
+	"gorm.io/gorm"
+
+	"github.com/marcbran/versource/internal"
+)
+
+type GormComponentRepo struct {
+	db *gorm.DB
+}
+
+func NewGormComponentRepo(db *gorm.DB) *GormComponentRepo {
+	return &GormComponentRepo{db: db}
+}
+
+func (r *GormComponentRepo) GetComponent(ctx context.Context, componentID uint) (*internal.Component, error) {
+	db := getTxOrDb(ctx, r.db)
+	var component internal.Component
+	err := db.WithContext(ctx).Where("id = ?", componentID).First(&component).Error
+	if err != nil {
+		return nil, fmt.Errorf("failed to get component: %w", err)
+	}
+	return &component, nil
+}
+
+func (r *GormComponentRepo) CreateComponent(ctx context.Context, component *internal.Component) error {
+	db := getTxOrDb(ctx, r.db)
+	err := db.WithContext(ctx).Create(component).Error
+	if err != nil {
+		return fmt.Errorf("failed to create component: %w", err)
+	}
+	return nil
+}
+
+func (r *GormComponentRepo) UpdateComponent(ctx context.Context, component *internal.Component) error {
+	db := getTxOrDb(ctx, r.db)
+	err := db.WithContext(ctx).Save(component).Error
+	if err != nil {
+		return fmt.Errorf("failed to update component: %w", err)
+	}
+	return nil
+}
